day02a -- much more complicated but "cleaner"
Creates an actual pad that knows the relative directions.
Pad buttons are vectors w/ a value and u/d/l/r references.
Edges point to themselves, so you don't have to check anything.

The hardest part is wiring up the pad, which is made a little easier
with transpose, each-slice, and in-subs