Optimize 19a for racket.
Goes from ~0.72s to 0.53s.

This was a pretty hefty overhaul without changing the overall
architecture (in rough order):

+ switch setters to static compilation = ~2-4 ms    (~0.75s -> ~0.72s)
+ mutable structs                      = ~150-200ms (~0.70s -> ~0.55s)
+ reg struct -> vector                 = ~0ms (possibly -5ms)
+ inst struct -> vector                = ~0ms (~0.59s)
+ currying reg-[gs]etter               = ~150ms ? (~0.45s)

Overall, it would have probably been fine to switch to mutable structs
and stop there, but what I really wanted was to have the function
stored in the instruction do ALL the work:

+ Access slots (or immediate values), via static access
+ Compute the result, as needed
+ Store the result, via static access

It turns out that this didn't make THAT much of a difference, but that
might be because of the length of the current problem. Maybe I could
figure out the input enough to make it run longer and test that theory.